partially integrated mockup
mockup has been integrated to the login and signup pages+ the my cases page.
home page has the location button and database has location.
orginisations page is compleately not done. map page is running but is not integrated and it gets userslocation but cannot get police stations and such and is not integrated. the cases do not check the user im working on that.